requirements.txt
asgiref==3.8.1
click==8.1.7
cmake==3.30.2
colorama==0.4.6
dj-database-url==2.2.0
Django==5.1
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
dlib==19.24.99
face-recognition==1.3.0
face_recognition_models==0.3.0
gunicorn==23.0.0
numpy==2.0.1
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
